requirements.txt
# MQTT client
paho-mqtt>=2.0.0

# Timezone handling
pytz>=2024.1

# Environment variable configuration
python-dotenv>=1.0.0

# P2: Config hot reload
watchdog>=4.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=1.0.0
